requirements_dev.txt
astroid==2.2.5
Django==2.2.1
django-debug-toolbar==1.11
django-environ==0.4.5
django-simple-menu==1.2.1
django-tables2==2.0.6
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
wrapt==1.11.1
